requirements.txt
fastapi==0.111.0
uvicorn==0.30.1
pydantic==2.7.1
openai==1.30.1
sentence-transformers==3.0.0
faiss-cpu==1.8.0
numpy==1.26.4
python-dotenv==1.0.1
httpx==0.27.0
